requirements.txt
Flask
Flask-WTF
WTForms
python-dotenv
google-generativeai
openai
